requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.27
langchain-openai==0.2.13
langchain-text-splitters==0.3.4
langsmith==0.2.4
marshmallow==3.23.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.0
openai==1.58.1
orjson==3.10.12
packaging==24.2
postgrest==0.18.0
propcache==0.2.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
PyYAML==6.0.2
realtime==2.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.18.3
